requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
azure-core==1.26.3
azure-storage-blob==12.14.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.2
cycler==0.12.1
Django==4.2.1
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==23.5
django-jazzmin==2.6.1
django-redis==5.4.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docx==0.2.4
docx2txt==0.8
drf-yasg==1.21.7
fonttools==4.53.1
fpdf==1.7.2
gunicorn==21.2.0
idna==3.6
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflection==0.5.1
isodate==0.6.1
kiwisolver==1.4.5
lxml==5.1.0
Markdown==3.5.2
matplotlib==3.9.1
msrest==0.7.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycparser==2.21
pygame==2.5.2
PyJWT==2.8.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyobjc==10.3.1
pyobjc-core==10.3.1
pyobjc-framework-Accessibility==10.3.1
pyobjc-framework-Accounts==10.3.1
pyobjc-framework-AddressBook==10.3.1
pyobjc-framework-AdServices==10.3.1
pyobjc-framework-AdSupport==10.3.1
pyobjc-framework-AppleScriptKit==10.3.1
pyobjc-framework-AppleScriptObjC==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-AppTrackingTransparency==10.3.1
pyobjc-framework-AudioVideoBridging==10.3.1
pyobjc-framework-AuthenticationServices==10.3.1
pyobjc-framework-AutomaticAssessmentConfiguration==10.3.1
pyobjc-framework-Automator==10.3.1
pyobjc-framework-AVFoundation==10.3.1
pyobjc-framework-AVKit==10.3.1
pyobjc-framework-AVRouting==10.3.1
pyobjc-framework-BackgroundAssets==10.3.1
pyobjc-framework-BrowserEngineKit==10.3.1
pyobjc-framework-BusinessChat==10.3.1
pyobjc-framework-CalendarStore==10.3.1
pyobjc-framework-CallKit==10.3.1
pyobjc-framework-CFNetwork==10.3.1
pyobjc-framework-Cinematic==10.3.1
pyobjc-framework-ClassKit==10.3.1
pyobjc-framework-CloudKit==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Collaboration==10.3.1
pyobjc-framework-ColorSync==10.3.1
pyobjc-framework-Contacts==10.3.1
pyobjc-framework-ContactsUI==10.3.1
pyobjc-framework-CoreAudio==10.3.1
pyobjc-framework-CoreAudioKit==10.3.1
pyobjc-framework-CoreBluetooth==10.3.1
pyobjc-framework-CoreData==10.3.1
pyobjc-framework-CoreHaptics==10.3.1
pyobjc-framework-CoreLocation==10.3.1
pyobjc-framework-CoreMedia==10.3.1
pyobjc-framework-CoreMediaIO==10.3.1
pyobjc-framework-CoreMIDI==10.3.1
pyobjc-framework-CoreML==10.3.1
pyobjc-framework-CoreMotion==10.3.1
pyobjc-framework-CoreServices==10.3.1
pyobjc-framework-CoreSpotlight==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-CoreWLAN==10.3.1
pyobjc-framework-CryptoTokenKit==10.3.1
pyobjc-framework-DataDetection==10.3.1
pyobjc-framework-DeviceCheck==10.3.1
pyobjc-framework-DictionaryServices==10.3.1
pyobjc-framework-DiscRecording==10.3.1
pyobjc-framework-DiscRecordingUI==10.3.1
pyobjc-framework-DiskArbitration==10.3.1
pyobjc-framework-DVDPlayback==10.3.1
pyobjc-framework-EventKit==10.3.1
pyobjc-framework-ExceptionHandling==10.3.1
pyobjc-framework-ExecutionPolicy==10.3.1
pyobjc-framework-ExtensionKit==10.3.1
pyobjc-framework-ExternalAccessory==10.3.1
pyobjc-framework-FileProvider==10.3.1
pyobjc-framework-FileProviderUI==10.3.1
pyobjc-framework-FinderSync==10.3.1
pyobjc-framework-FSEvents==10.3.1
pyobjc-framework-GameCenter==10.3.1
pyobjc-framework-GameController==10.3.1
pyobjc-framework-GameKit==10.3.1
pyobjc-framework-GameplayKit==10.3.1
pyobjc-framework-HealthKit==10.3.1
pyobjc-framework-ImageCaptureCore==10.3.1
pyobjc-framework-InputMethodKit==10.3.1
pyobjc-framework-InstallerPlugins==10.3.1
pyobjc-framework-InstantMessage==10.3.1
pyobjc-framework-Intents==10.3.1
pyobjc-framework-IntentsUI==10.3.1
pyobjc-framework-IOBluetooth==10.3.1
pyobjc-framework-IOBluetoothUI==10.3.1
pyobjc-framework-IOSurface==10.3.1
pyobjc-framework-iTunesLibrary==10.3.1
pyobjc-framework-KernelManagement==10.3.1
pyobjc-framework-LatentSemanticMapping==10.3.1
pyobjc-framework-LaunchServices==10.3.1
pyobjc-framework-libdispatch==10.3.1
pyobjc-framework-libxpc==10.3.1
pyobjc-framework-LinkPresentation==10.3.1
pyobjc-framework-LocalAuthentication==10.3.1
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.3.1
pyobjc-framework-MailKit==10.3.1
pyobjc-framework-MapKit==10.3.1
pyobjc-framework-MediaAccessibility==10.3.1
pyobjc-framework-MediaLibrary==10.3.1
pyobjc-framework-MediaPlayer==10.3.1
pyobjc-framework-MediaToolbox==10.3.1
pyobjc-framework-Metal==10.3.1
pyobjc-framework-MetalFX==10.3.1
pyobjc-framework-MetalKit==10.3.1
pyobjc-framework-MetalPerformanceShaders==10.3.1
pyobjc-framework-MetalPerformanceShadersGraph==10.3.1
pyobjc-framework-MetricKit==10.3.1
pyobjc-framework-MLCompute==10.3.1
pyobjc-framework-ModelIO==10.3.1
pyobjc-framework-MultipeerConnectivity==10.3.1
pyobjc-framework-NaturalLanguage==10.3.1
pyobjc-framework-NetFS==10.3.1
pyobjc-framework-Network==10.3.1
pyobjc-framework-NetworkExtension==10.3.1
pyobjc-framework-NotificationCenter==10.3.1
pyobjc-framework-OpenDirectory==10.3.1
pyobjc-framework-OSAKit==10.3.1
pyobjc-framework-OSLog==10.3.1
pyobjc-framework-PassKit==10.3.1
pyobjc-framework-PencilKit==10.3.1
pyobjc-framework-PHASE==10.3.1
pyobjc-framework-Photos==10.3.1
pyobjc-framework-PhotosUI==10.3.1
pyobjc-framework-PreferencePanes==10.3.1
pyobjc-framework-PushKit==10.3.1
pyobjc-framework-Quartz==10.3.1
pyobjc-framework-QuickLookThumbnailing==10.3.1
pyobjc-framework-ReplayKit==10.3.1
pyobjc-framework-SafariServices==10.3.1
pyobjc-framework-SafetyKit==10.3.1
pyobjc-framework-SceneKit==10.3.1
pyobjc-framework-ScreenCaptureKit==10.3.1
pyobjc-framework-ScreenSaver==10.3.1
pyobjc-framework-ScreenTime==10.3.1
pyobjc-framework-ScriptingBridge==10.3.1
pyobjc-framework-SearchKit==10.3.1
pyobjc-framework-Security==10.3.1
pyobjc-framework-SecurityFoundation==10.3.1
pyobjc-framework-SecurityInterface==10.3.1
pyobjc-framework-SensitiveContentAnalysis==10.3.1
pyobjc-framework-ServiceManagement==10.3.1
pyobjc-framework-SharedWithYou==10.3.1
pyobjc-framework-SharedWithYouCore==10.3.1
pyobjc-framework-ShazamKit==10.3.1
pyobjc-framework-Social==10.3.1
pyobjc-framework-SoundAnalysis==10.3.1
pyobjc-framework-Speech==10.3.1
pyobjc-framework-SpriteKit==10.3.1
pyobjc-framework-StoreKit==10.3.1
pyobjc-framework-Symbols==10.3.1
pyobjc-framework-SyncServices==10.3.1
pyobjc-framework-SystemConfiguration==10.3.1
pyobjc-framework-SystemExtensions==10.3.1
pyobjc-framework-ThreadNetwork==10.3.1
pyobjc-framework-UniformTypeIdentifiers==10.3.1
pyobjc-framework-UserNotifications==10.3.1
pyobjc-framework-UserNotificationsUI==10.3.1
pyobjc-framework-VideoSubscriberAccount==10.3.1
pyobjc-framework-VideoToolbox==10.3.1
pyobjc-framework-Virtualization==10.3.1
pyobjc-framework-Vision==10.3.1
pyobjc-framework-WebKit==10.3.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.0.2
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.3
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tk==0.1.0
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.19.2
